token-perustainen kirjautuminen

+++ osa04/blogilista/controllers/blogs.js
@@ -1,27 +1,43 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require ('../models/user')
+const jwt = require('jsonwebtoken')
 
-blogsRouter.get('/', (request, response) => {
-  Blog
-    .find({})
-    .then(blogs => {
-      response.json(blogs)
-    })
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
+  response.json(blogs)
 })
 
-blogsRouter.post('/', (request, response, next) => {
-  const blog = new Blog(request.body)
-  blog
-    .save()
-    .then(result => {
-      response.status(201).json(result)
-    })
-    .catch(error => next(error))
+
+blogsRouter.post('/', async (request, response) => {
+  const { author, title, url, likes } = request.body
+  const user = request.user
+  if (!request.user) {
+    return response.status(401)
+  }
+  const blog = new Blog({
+    author,
+    title,
+    url,
+    likes: likes,
+    user: user._id
+  })
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if (user && user._id.toString() === blog.user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).end()
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -30,7 +46,7 @@ blogsRouter.put('/:id', async (request, response) => {
     author: body.author,
     title: body.title,
     url: body.url,
-    likes: Number(body.likes)
+    likes: Number(body.likes),
   }
   const updatedBlog = await Blog.findByIdAndUpdate(
     request.params.id,
